Change TokenVerifier to an interface
