+++ data-plane/core/src/main/java/dev/knative/eventing/kafka/broker/core/oidc/TokenVerifier.java
@@ -16,62 +16,10 @@
 package dev.knative.eventing.kafka.broker.core.oidc;
 
 import io.vertx.core.Future;
-import io.vertx.core.Vertx;
 import io.vertx.core.http.HttpServerRequest;
 import org.jose4j.jwt.JwtClaims;
-import org.jose4j.jwt.consumer.InvalidJwtException;
-import org.jose4j.jwt.consumer.JwtConsumer;
-import org.jose4j.jwt.consumer.JwtConsumerBuilder;
-import org.jose4j.jwt.consumer.JwtContext;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-public class TokenVerifier {
-
-    private static final Logger logger = LoggerFactory.getLogger(TokenVerifier.class);
-
-    private final Vertx vertx;
-
-    private final OIDCDiscoveryConfig oidcDiscoveryConfig;
-
-    public TokenVerifier(Vertx vertx, OIDCDiscoveryConfig oidcDiscoveryConfig) {
-        this.vertx = vertx;
-        this.oidcDiscoveryConfig = oidcDiscoveryConfig;
-    }
-
-    public Future<JwtClaims> verify(String token, String expectedAudience) {
-        return this.vertx.<JwtClaims>executeBlocking(promise -> {
-            // execute blocking, as jose .process() is blocking
-
-            JwtConsumer jwtConsumer = new JwtConsumerBuilder()
-                    .setVerificationKeyResolver(this.oidcDiscoveryConfig.getJwksVerificationKeyResolver())
-                    .setExpectedAudience(expectedAudience)
-                    .setExpectedIssuer(this.oidcDiscoveryConfig.getIssuer())
-                    .build();
-
-            try {
-                JwtContext jwtContext = jwtConsumer.process(token);
-
-                promise.complete(jwtContext.getJwtClaims());
-            } catch (InvalidJwtException e) {
-                promise.fail(e);
-            }
-        });
-    }
-
-    public Future<JwtClaims> verify(final HttpServerRequest request, String expectedAudience) {
-        String authHeader = request.getHeader("Authorization");
-        if (authHeader == null || authHeader.isEmpty()) {
-            return Future.failedFuture("Request didn't contain Authorization header"); // change to exception
-        }
-
-        if (!authHeader.startsWith("Bearer ") && authHeader.length() <= "Bearer ".length()) {
-            return Future.failedFuture("Authorization header didn't contain Bearer token"); // change to exception
-        }
-
-        String token = authHeader.substring("Bearer ".length());
-
-        request.pause();
-        return verify(token, expectedAudience).onSuccess(v -> request.resume());
-    }
+public interface TokenVerifier {
+    Future<JwtClaims> verify(String token, String expectedAudience);
+    Future<JwtClaims> verify(HttpServerRequest request, String expectedAudience);
 }
